Allow command tags up to 64 bits - not required, backward compatible

+++ src/radio.h
@@ -313,7 +313,7 @@ struct channel {
 
   struct {
     uint64_t packets_in;
-    uint32_t tag;
+    uint64_t tag;               // arbitrary value computed by client and sent in status responses
     pthread_mutex_t lock;       // Protect statistics during updates and reads
     uint64_t blocks_since_poll; // Used for averaging signal levels
     int global_timer;
